Move shareWithFriend to ViewContainer, pass as prop, add option to PinEditButton
