+++ app/components/PinListItem.js
@@ -9,8 +9,6 @@ import React, {
 } from 'react-native';
 
 import { Actions } from 'react-native-router-flux';
-import FriendList from './FriendList';
-import { ref } from '../lib/db/db';
 import Location from '../lib/orientation/locationMath';
 
 
@@ -21,7 +19,7 @@ export default class PinListItem extends Component {
   }
 
   touchOptions() {
-    const { pin, friends, deletePin, setTarget, redraw } = this.props;
+    const { pin, friends, deletePin, setTarget, redraw, shareWithFriend } = this.props;
     AlertIOS.prompt(
         pin.title,
         '('+pin.longitude + ', ' + pin.latitude + ')',
@@ -35,7 +33,7 @@ export default class PinListItem extends Component {
         },
         {
           text: 'Share',
-          onPress: () => { Actions.friends({ onPress: this.shareWithFriend.bind( this, pin ), friends: friends }) },
+          onPress: () => { Actions.friends({ onPress: shareWithFriend.bind( null, pin ), friends: friends }) },
         },
         {
           text: 'Set Target',
@@ -52,35 +50,6 @@ export default class PinListItem extends Component {
         }],
         'plain-text'
       );
-  }
-
-  shareWithFriend( pin, friend ) {
-    const { user } = this.props;
-    if( typeof user.id !== 'string' ) {
-      console.log( 'shareWithFriend: user id must be a string' );
-      return null;
-    }
-    if( typeof friend.id !== 'string' ) {
-      console.log( 'shareWithFriend: friend id must be a string' );
-      return null;
-    }
-    if( typeof pin !== 'object' ) {
-      console.log( 'shareWithFriend: pin must be an object' );
-      return null;
-    }
-    if( typeof pin.id !== 'string' ) {
-      console.log( 'shareWithFriend: pin id must be a string' );
-      return null;
-    }
-    // Make a copy of the pin
-
-    var pinCopy = Object.assign({}, {alertedYet: false} ,pin);
-    // Set pin.friend to the userID of the person sending the pin
-    pinCopy.friend = user;
-    // Post the pin to the friend's firebase.
-    var friendPin = ref.child( friend.id ).child( 'pins' ).child( pin.id );
-    friendPin.set( pinCopy );
-    return true;
   }
 
   editTitle(value) {
