+++ app/components/ViewContainer.js
@@ -4,7 +4,7 @@ import Map from './Map';
 import DropNewPinButton from '../containers/container_dropNewPin';
 import PinList from './PinList';
 import Button from 'react-native-button';
-import { myCurrLoc } from '../lib/db/db';
+import { ref, myCurrLoc } from '../lib/db/db';
 
 const styles = StyleSheet.create({
   ViewMenu: {
@@ -42,6 +42,35 @@ export default class ViewContainer extends Component {
     this.setState({ view });
   }
 
+  shareWithFriend( pin, friend ) {
+    const { user } = this.props;
+    if( typeof user.id !== 'string' ) {
+      console.log( 'shareWithFriend: user id must be a string' );
+      return null;
+    }
+    if( typeof friend.id !== 'string' ) {
+      console.log( 'shareWithFriend: friend id must be a string' );
+      return null;
+    }
+    if( typeof pin !== 'object' ) {
+      console.log( 'shareWithFriend: pin must be an object' );
+      return null;
+    }
+    if( typeof pin.id !== 'string' ) {
+      console.log( 'shareWithFriend: pin id must be a string' );
+      return null;
+    }
+    // Make a copy of the pin
+
+    var pinCopy = Object.assign({}, {alertedYet: false} ,pin);
+    // Set pin.friend to the userID of the person sending the pin
+    pinCopy.friend = user;
+    // Post the pin to the friend's firebase.
+    var friendPin = ref.child( friend.id ).child( 'pins' ).child( pin.id );
+    friendPin.set( pinCopy );
+    return true;
+  }
+
   render() {
     const { pins, recent, friends, user, targetPin, getLocationToSave, updatePins, updateRecent, deletePin, setTarget, clearTarget } = this.props;
 
@@ -50,6 +79,7 @@ export default class ViewContainer extends Component {
       <View style={{flex: 1}}>
       { this.state.view === 'ar' ? <AR pins={pins} targetPin={targetPin} /> : void 0 }
       { this.state.view === 'map' ? <Map
+          shareWithFriend={this.shareWithFriend.bind(this)}
           getLocationToSave={getLocationToSave}
           // initialLoc={this.state.initialLoc}
           pins = {pins}
@@ -63,6 +93,7 @@ export default class ViewContainer extends Component {
           clearTarget={clearTarget}
         /> : void 0 }
       { this.state.view === 'list' ? <PinList
+          shareWithFriend={this.shareWithFriend.bind(this)}
           deletePin={deletePin}
           updatePins={updatePins}
           updateRecent={updateRecent}
